Web service client stub regenerated. getOid changed for getId calls. refresh-ws.sh script updated

+++ client/trunk/Communications/src/org/inventory/communications/wsclient/RemoteActor.java
@@ -16,7 +16,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}long"/>
  *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="type" type="{http://www.w3.org/2001/XMLSchema}int"/>
  *       &lt;/sequence>
@@ -35,7 +35,7 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class RemoteActor {
 
-    protected int id;
+    protected long id;
     protected String name;
     protected int type;
 
@@ -43,7 +43,7 @@ public class RemoteActor {
      * Gets the value of the id property.
      * 
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -51,7 +51,7 @@ public class RemoteActor {
      * Sets the value of the id property.
      * 
      */
-    public void setId(int value) {
+    public void setId(long value) {
         this.id = value;
     }
